user permissions and administration

+++ app.js
@@ -20,7 +20,7 @@ db.run("CREATE TABLE IF NOT EXISTS users ( id INTEGER PRIMARY KEY AUTOINCREMENT,
       if(err) { console.log(err); }
       if(!row) {
         bcrypt.hash(settings.app.defaultpass, saltRounds, function(err, hash) {
-          db.run("INSERT INTO users ( user, password, permissions ) VALUES( ?, ?, 'super' )", settings.app.defaultuser, hash, function(err) {
+          db.run("INSERT INTO users ( user, password, permissions ) VALUES( ?, ?, 'admin' )", settings.app.defaultuser, hash, function(err) {
             if(err) { console.log(err); }
           });
         });
@@ -31,21 +31,22 @@ db.run("CREATE TABLE IF NOT EXISTS users ( id INTEGER PRIMARY KEY AUTOINCREMENT,
 // db functions
 var createUser = function(req, res) {
   var user = req.body.user;
+  var permissions = req.body.permissions;
   db.get('SELECT user FROM users WHERE user = ?', user, function(err, row) {
     // if(err)
     if(row) {
-      req.flash('errorMessage', 'There is already a user with that name.');
-      // res.redirect('/admin/users');
+      req.flash('errorMessage', " there is already a user with that name");
+      res.redirect('/admin/users');
     } else {
       bcrypt.hash(req.body.password, saltRounds, function(err, hash) {
-        db.run('INSERT INTO users ( user, password ) VALUES( ?, ? )', user, hash, function(err) {
+        db.run('INSERT INTO users ( user, password, permissions ) VALUES( ?, ?, ? )', user, hash, permissions, function(err) {
           // if(err)
           if(this.lastID) {
-            req.flash('successMessage', 'User created!');
-            // res.redirect('/admin/users');
+            req.flash('successMessage', " user created");
+            res.redirect('/admin/users');
           } else {
-            req.flash('errorMessage', 'Apologies, it seems something went wrong.');
-            // res.redirect('/admin/users');
+            req.flash('errorMessage', " something went wrong");
+            res.redirect('/admin/users');
           }
         });
       });
@@ -54,19 +55,65 @@ var createUser = function(req, res) {
 }
 
 var deleteUser = function(req, res) {
-  if(req.user.user === req.body.user) {
-    req.flash('errorMessage', "You can't delete yourself.");
-    // res.redirect('/admin/users');
+  if(req.user.id == req.body.id) {
+    req.flash('errorMessage', " you can't delete yourself");
+    res.redirect('/admin/users');
   } else {
-    db.run('DELETE FROM users WHERE user = ?', req.body.user, function(err) {
+    db.run('DELETE FROM users WHERE id = ?', req.body.id, function(err) {
       if(err) {
         //...
       } else {
-        req.flash('successMessage', 'User deleted!');
-        // res.redirect('/admin/users');
+        req.flash('successMessage', " user deleted");
+        res.redirect('/admin/users');
       }
     });
   }
+}
+
+
+var editUser = function(req, res) {
+  var runquery = false;
+  var query = "UPDATE users SET ";
+  if(req.body.password.length == 0) {
+    req.flash('successMessage', " password not changed");
+  } else {
+    bcrypt.hash(req.body.password, saltRounds, function(err, hash) {
+      // if(err)
+      query += "password = " + hash + ", " ;
+      runquery = true;
+    });
+  }
+  if(req.user.id == req.body.id && req.body.permissions != "admin") {
+    req.flash('errorMessage', " you can't remove your own admin status");
+  } else {
+    query += "permissions = '" + req.body.permissions + "'";
+    runquery = true;
+  }
+  query += " WHERE id = " + req.body.id;
+  if(runquery == true) {
+    db.run(query, function(err) {
+      if(err) {
+        //...
+        console.log(err);
+        req.flash('errorMessage', " something went wrong");
+        res.redirect('/admin/users');
+      } else {
+        req.flash('successMessage', " user updated");
+        res.redirect('/admin/users');
+      }
+    });
+  } else {
+    req.flash('errorMessage', " something went wrong");
+    res.redirect('/admin/users');
+  }
+
+}
+
+var listUsers = function(cb) {
+  db.all('SELECT id, user, permissions FROM users', function(err, rows) {
+    // if(err)
+    cb(rows);
+  });
 }
 
 // setting up user authentication
@@ -87,7 +134,7 @@ passport.serializeUser(function(user, done) {
 });
 
 passport.deserializeUser(function(id, done) {
-  db.get('SELECT id, user FROM users WHERE id = ?', id, function(err, row) {
+  db.get('SELECT id, user, permissions FROM users WHERE id = ?', id, function(err, row) {
     if (!row) { return done(null, false); }
     return done(null, row);
   });
@@ -120,7 +167,17 @@ app.use(flash());
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.engine('handlebars', exphbs({defaultLayout: 'main'}));
+app.engine('handlebars', exphbs({
+  defaultLayout: 'main',
+  helpers: {
+    eq: function(v1,v2,options) {
+			if (v1 && v2 && v1.toString() === v2.toString()) {
+				return options.fn(this);
+			}
+			return options.inverse(this);
+		}
+  }
+}));
 app.set('view engine', 'handlebars');
 
 app.use(express.static('public'));
@@ -144,20 +201,35 @@ app.get('/', function(req, res) {
 		});
 });
 
-app.post('/user', function(req, res) {
-  if (req.user) {
+app.get('/admin/users', function(req, res) {
+  if(req.user && req.user.permissions == "admin") {
+    listUsers(function(result) {
+      res.render('users',{
+        user:req.user,
+        users: result,
+        error:req.flash("errorMessage"),
+        success:req.flash("successMessage")
+      });
+    });
+  } else {
+    res.redirect('/');
+  }
+});
+
+app.post('/admin/user', function(req, res) {
+  if (req.user && req.user.permissions == "admin") {
     switch(req.body["_method"]) {
       case "DELETE":
         deleteUser(req, res);
       break;
       case "PUT":
-        // ...
+        editUser(req, res);
       break;
       default:
         createUser(req, res);
       break;
     }
-  } else { res.redirect('/admin'); }
+  } else { res.redirect('/admin/users'); }
 });
 
 app.listen(settings.app.port, function() {
